+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/rolling/Availability.java
@@ -1,0 +1,49 @@
+package io.strimzi.operator.cluster.operator.resource.rolling;
+
+import java.util.Map;
+
+public class Availability {
+
+
+    protected static boolean anyReplicaWouldBeUnderReplicated(KafkaNode kafkaNode,
+                                                              Map<String, Integer> minIsrByTopic) {
+        for (var replica : kafkaNode.replicas()) {
+            var topicName = replica.topicName();
+            Integer minIsr = minIsrByTopic.get(topicName);
+            if (wouldBeUnderReplicated(minIsr, replica)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+
+    private static boolean wouldBeUnderReplicated(Integer minIsr, Replica replica) {
+        final boolean wouldByUnderReplicated;
+        if (minIsr == null) {
+            // if topic doesn't have minISR then it's fine
+            wouldByUnderReplicated = false;
+        } else {
+            // else topic has minISR
+            // compute spare = size(ISR) - minISR
+            int sizeIsr = replica.isrSize();
+            int spare = sizeIsr - minIsr;
+            if (spare > 0) {
+                // if (spare > 0) then we can restart the broker hosting this replica
+                // without the topic being under-replicated
+                wouldByUnderReplicated = false;
+            } else if (spare == 0) {
+                // if we restart this broker this replica would be under-replicated if it's currently in the ISR
+                // if it's not in the ISR then restarting the server won't make a difference
+                wouldByUnderReplicated = replica.isInIsr();
+            } else {
+                // this partition is already under-replicated
+                // if it's not in the ISR then restarting the server won't make a difference
+                // but in this case since it's already under-replicated let's
+                // not possible prolong the time to this server rejoining the ISR
+                wouldByUnderReplicated = true;
+            }
+        }
+        return wouldByUnderReplicated;
+    }
+}
