+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/rolling/Batching.java
@@ -14,7 +14,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
 public class Batching {
@@ -172,7 +171,7 @@ public class Batching {
         for (var cell : cells) {
             List<Set<KafkaNode>> availBatches = new ArrayList<>();
             for (var kafkaNode : cell) {
-                if (affectsAvailability(kafkaNode, minIsrByTopic)) {
+                if (!Availability.anyReplicaWouldBeUnderReplicated(kafkaNode, minIsrByTopic)) {
                     LOGGER.debugCr(reconciliation, "No replicas of node {} will be unavailable => add to batch",
                             kafkaNode.id());
                     var currentBatch = availBatches.isEmpty() ? null : availBatches.get(availBatches.size() - 1);
@@ -257,46 +256,5 @@ public class Batching {
     }
 
 
-    protected static boolean affectsAvailability(KafkaNode kafkaNode,
-                                               Map<String, Integer> minIsrByTopic) {
-        for (var replica : kafkaNode.replicas()) {
-            var topicName = replica.topicName();
-            Integer minIsr = minIsrByTopic.get(topicName);
-            if (wouldBeUnderReplicated(minIsr, replica)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-
-    private static boolean wouldBeUnderReplicated(Integer minIsr, Replica replica) {
-        final boolean wouldByUnderReplicated;
-        if (minIsr == null) {
-            // if topic doesn't have minISR then it's fine
-            wouldByUnderReplicated = false;
-        } else {
-            // else topic has minISR
-            // compute spare = size(ISR) - minISR
-            int sizeIsr = replica.isrSize();
-            int spare = sizeIsr - minIsr;
-            if (spare > 0) {
-                // if (spare > 0) then we can restart the broker hosting this replica
-                // without the topic being under-replicated
-                wouldByUnderReplicated = false;
-            } else if (spare == 0) {
-                // if we restart this broker this replica would be under-replicated if it's currently in the ISR
-                // if it's not in the ISR then restarting the server won't make a difference
-                wouldByUnderReplicated = replica.isInIsr();
-            } else {
-                // this partition is already under-replicated
-                // if it's not in the ISR then restarting the server won't make a difference
-                // but in this case since it's already under-replicated let's
-                // not possible prolong the time to this server rejoining the ISR
-                wouldByUnderReplicated = true;
-            }
-        }
-        return wouldByUnderReplicated;
-    }
 
 }
